requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pip==23.2.1
pymongo==4.5.0
pydantic==2.3.0
email-validator==2.0.0.post2 
pytest==7.4.0
httpx==0.24.1
ipython==8.14.0
motor==3.3.0
mysql-connector-python==8.2.0